Struct slog_envlogger::LogBuilder

pub struct LogBuilder<T: Drain> { /* fields omitted */ }

LogBuilder acts as builder for initializing the EnvLogger. It can be used change the enviromental variable used to provide the logging directives and also set the default log level filter.

Methods

impl<T: Drain> LogBuilder<T>

fn new(d: T) -> Self

Initializes the log builder with defaults

fn filter(self, module: Option<&str>, level: FilterLevel) -> Self

Adds filters to the logger

The given module (if any) will log at most the specified level provided. If no module is provided then the filter will apply to all log messages.

fn parse(self, filters: &str) -> Self

Parses the directives string in the same form as the RUST_LOG environment variable.

See the module documentation for more details.

fn build(self) -> EnvLogger<T>

Build an env logger.